Combine runtime docstring + signature modification into a single decorator
* Also combine docstrings with template functions
* Add unit tests for new documentation utils
