+++ pyinaturalist/api_requests.py
@@ -68,6 +68,7 @@ def request(
     headers = headers or {}
     headers["Accept"] = "application/json"
     headers["User-Agent"] = user_agent or pyinaturalist.user_agent
+
     if access_token:
         headers["Authorization"] = "Bearer %s" % access_token
     params = preprocess_request_params(params)
